PIM-3832: move classes from Connector Bundle to Component
