+++ src/Pim/Bundle/ConnectorBundle/Resources/config/processors.yml
@@ -1,8 +1,8 @@
 parameters:
-    pim_base_connector.processor.denormalization.attribute_option.class:                  Pim\Bundle\ConnectorBundle\Processor\Denormalization\AttributeOptionProcessor
+    pim_base_connector.processor.denormalization.attribute_option.class:                  Pim\Component\Connector\Processor\Denormalization\AttributeOptionProcessor
 
-    pim_base_connector.denormalization.array_converter.flat.attribute_option.class:       Pim\Bundle\ConnectorBundle\Processor\Denormalization\ArrayConverter\Flat\AttributeOptionToStandardConverter
-    pim_base_connector.denormalization.array_converter.structured.attribute_option.class: Pim\Bundle\ConnectorBundle\Processor\Denormalization\ArrayConverter\Structured\AttributeOptionToStandardConverter
+    pim_base_connector.denormalization.array_converter.flat.attribute_option.class:       Pim\Component\Connector\Processor\Denormalization\ArrayConverter\Flat\AttributeOptionToStandardConverter
+    pim_base_connector.denormalization.array_converter.structured.attribute_option.class: Pim\Component\Connector\Processor\Denormalization\ArrayConverter\Structured\AttributeOptionToStandardConverter
 
 services:
     # Denormalization - Processors
