+++ src/Pim/Component/Connector/Processor/Denormalization/ArrayConverter/Flat/AttributeOptionToStandardConverter.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Pim\Bundle\ConnectorBundle\Processor\Denormalization\ArrayConverter\Flat;
+namespace Pim\Component\Connector\Processor\Denormalization\ArrayConverter\Flat;
 
-use Pim\Bundle\ConnectorBundle\Exception\ArrayConversionException;
-use Pim\Bundle\ConnectorBundle\Processor\Denormalization\ArrayConverter\StandardArrayConverterInterface;
 use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
+use Pim\Component\Connector\Exception\ArrayConversionException;
+use Pim\Component\Connector\Processor\Denormalization\ArrayConverter\StandardArrayConverterInterface;
 
 /**
  * Convert flat format to standard format for attribute option
